+++ CHANGES.md
@@ -6,6 +6,8 @@ Changes
 
 ### Unreleased
 
+* 2020-08-11 - Added setting to be able to show a hint for a unrestricted self enrolment in a visible course.
+               NOTE: Capability theme/boost_campus:viewhintcourseselfenrol introduced.
 * 2020-08-07 - Added rule to show description border for all users for restricted activities.
 * 2020-08-07 - Added check for empty string to regular expression for time controlled info banner settings.
 * 2020-08-05 - Added settings to be able to show a banner with information on selected pages.